+++ ff_standard_lib/src/consolidators/heikinashi.rs
@@ -248,7 +248,7 @@ impl HeikinAshiConsolidator
     }
     
 
-    pub(crate) fn index(&self, index: usize) -> Option<BaseDataEnum> {
+    pub(crate) fn index(&self, index: u64) -> Option<BaseDataEnum> {
         match self.history.get(index) {
             Some(data) => Some(data.clone()),
             None => None,
