+++ ff_standard_lib/src/consolidators/renko.rs
@@ -115,7 +115,7 @@ impl RenkoConsolidator
     }
 
 
-    pub(crate) fn index(&self, index: usize) -> Option<BaseDataEnum> {
+    pub(crate) fn index(&self, index: u64) -> Option<BaseDataEnum> {
         match self.history.get(index) {
             Some(data) => Some(data.clone()),
             None => None,
